NOISSUE Refactor VersionList class for improved clarity and ensure proper type casting in count method

+++ meshmc/launcher/meta/VersionList.cpp
@@ -69,7 +69,7 @@ namespace Meta
 	}
 	int VersionList::count() const
 	{
-		return m_versions.size();
+		return static_cast<int>(m_versions.size());
 	}
 
 	void VersionList::sortVersions()
